feat: 알림기능을 추가합니다.

+++ src/main/java/org/zerock/wego/domain/common/FavoriteVO.java
@@ -7,9 +7,9 @@ import lombok.Value;
 @Value
 public class FavoriteVO {
 	
-	private long userId;
+	private Integer userId;
 	private String targetGb;
-	private long targetCd;
+	private Integer targetCd;
 	private Timestamp modifiedDt;
 	private Character status;
